% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amatheme.R
\name{amatheme}
\alias{amatheme}
\title{American-Marketing-Association-ready theme for plots}
\usage{
amatheme()
}
\value{
A custom ggplot2 theme object.
}
\description{
Create a custom ggplot2 theme that complies with the American Marketing Association style.
This theme removes major and minor gridlines, panel borders, and adds black lines along the axes.
}
