% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_fig_ama.R
\name{export_fig_ama}
\alias{export_fig_ama}
\title{Function to export a figure with AMA theme}
\usage{
export_fig_ama(figure, filename, filepath)
}
\arguments{
\item{figure}{A ggplot2 object.}

\item{filename}{A character string specifying the filename.}

\item{filepath}{A character string specifying the directory to save the file.}
}
\description{
Function to export a figure with AMA theme
}
