% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tab_ama.R
\name{export_tab_ama}
\alias{export_tab_ama}
\title{Function to export a table}
\usage{
export_tab_ama(table, filename, filepath, caption)
}
\arguments{
\item{table}{A data frame or matrix.}

\item{filename}{A character string specifying the filename.}

\item{filepath}{A character string specifying the directory to save the file.}

\item{caption}{A character string specifying the caption for the table.}
}
\description{
Function to export a table
}
